+++ headers/Sphere.hpp
@@ -7,12 +7,12 @@
 class Sphere: public Hitable{
 public:
 	Vec3 center;
-	double radius;
+	float radius;
 	Material *material;
 
 	Sphere(){}
 
-	Sphere(Vec3 cen, double r, Material *mat){
+	Sphere(Vec3 cen, float r, Material *mat){
 		center = cen;
 		radius = r;
 		material = mat;
@@ -25,14 +25,14 @@ public:
 		return true;
 	}
 
-	virtual bool hit(const Ray& r, double t_min, double t_max, HitRecord& rec) const override{
+	virtual bool hit(const Ray& r, float t_min, float t_max, HitRecord& rec) const override{
 		Vec3 oc = r.origin() - center;
-		double a = Vec3::dot(r.direction(), r.direction());
-		double b = Vec3::dot(oc, r.direction());
-		double c = Vec3::dot(oc, oc) - radius*radius;
-		double delta = b*b - a*c;
+		float a = Vec3::dot(r.direction(), r.direction());
+		float b = Vec3::dot(oc, r.direction());
+		float c = Vec3::dot(oc, oc) - radius*radius;
+		float delta = b*b - a*c;
 		if (delta > 0){
-			double temp = (-b -sqrt(delta))/a;
+			float temp = (-b -sqrt(delta))/a;
 			if (temp > t_max || temp < t_min)
 				temp = (-b +sqrt(delta))/a;
 			if (temp < t_max && temp > t_min){
